place decompress code into seperate classes and files
